[style] return; -> return <바로이전수행코드>; 로 변경
[fix] get 서비스 결과값 체크후 에러 발생로직 제거
- 해당 서비스를 호출하는 부모 메서드에서 에러처리를 수행하도록 변경

+++ services/friend-service.js
@@ -139,10 +139,10 @@ async function deleteFriendList(userId, friendPkIdList) {
  * @returns {number} 삭제된 친구 수
  */
 async function deleteFriend(userId, friendPkId) {
-  const friend = await Friend.findByPk(friendPkId);
-  logger.debug(friend);
-
-  verifyUserFriend(friend, userId);
+  const friend = await getFriend(userId, friendPkId);
+  if (!friend) {
+    throw new BizError('친구 정보를 찾을 수 없습니다.');
+  }
 
   const result = await Friend.destroy({
     where: {
